+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - Dependency Updates
 - Omit undefined properties from geojson
+- Better geojson tests
 
 ## [4.13.1] - 2024-07-12